# [FINE] Faster report result index pages