--- a/app/controllers/application_controller.rb
+++ b/app/controllers/application_controller.rb
@@ -297,7 +297,7 @@ class ApplicationController < ActionController::Base
     @sb[:pages][:perpage] = settings(:perpage, :reports)
 
     rr = MiqReportResult.find(@sb[:pages][:rr_id])
-    @html = report_build_html_table(rr.report_results,
+    @html = report_build_html_table(rr.report,
                                     rr.html_rows(:page     => @sb[:pages][:current],
                                                  :per_page => @sb[:pages][:perpage]).join)