This is the Fine port of #4143.  Original PR message below.


* * *


Merge after https://github.com/ManageIQ/manageiq/pull/17590  (**NOTE:** Not relavent, this is already backported)

This makes both the `ChargebackController` and `ReportController::SavedReports` actions for viewing reports signficantly faster by:

* avoiding loading the `binary_blob` data entirely (avoids massive YAML deserialization).  This uses methods provided by https://github.com/ManageIQ/manageiq/pull/17590 to accomplish this.
* Fix button helpers to use `.exists?` instead of `.length > 0`, which prevents needing to fetch all of the report result row records, which were then thrown away and never used.


Benchmarks
----------

This is displaying a `MiqReportResult` index page that has 23k+ rows associated with it.  Showing the first page, paginated by 20 records.

**Before**

|   ms | queries | query (ms) |  rows |
| ---: |    ---: |       ---: |  ---: |
| 5701 |      20 |      178.9 | 47523 |
| 4475 |      20 |      161.2 | 47523 |
| 5070 |      20 |      151.8 | 47523 |
| 4976 |      20 |      140.1 | 47523 |
| 5189 |      20 |      141.4 | 47523 |


**After**

|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
|  462 |      19 |       55.8 |   30 |
|  234 |      19 |       53.5 |   30 |
|  272 |      19 |       51.0 |   30 |
|  234 |      19 |       46.2 |   30 |
|  283 |      19 |       55.8 |   30 |


Links
-----

* Main repo PR (~~Merge First!~~ Already done):  https://github.com/ManageIQ/manageiq/pull/17590
* https://bugzilla.redhat.com/show_bug.cgi?id=1594387


QA Steps
--------

I tested the did the following when running the `Before`/`After` benchmarks:

* Log in as an `admin` and go to `Cloud Intel` -> `Reports`
* Find a report that has a large number of report results, chargeback with some subtotals is prefered
* Select that report on the tree view, and select one of the report results table that is loaded in
* After the first paginated page has been returned, change the size of the page to 200 or something.  Make sure the switch between is also snappy.  (I toggled between page sizes with little to no wait time with all of the fixes in place).
* Make sure you can download using the `Download` toolbar button as well.